precommit: `terraform fmt` makes changes on disk
 ### Issues Addressed
Most of our pre-commit checks just make the changes that are rote,
obvious changes. Not `terraform fmt`! And there's no reason for it, it's
just someone decided to do it that way and the maintainers won't switch
it back. Comment thread:
https://github.com/gruntwork-io/pre-commit/issues/48

 ### Summary of Changes
I patched precommit to use the expected behavior. I included a shell
file to add the patch so it's easy to stay up-to-date with upstream.

 ### Test Plan
- Tested that precommit now automatically updates terraform files when
  using this branch for precommit.